fix(app): enable composer shortcuts on drafts (#33956)
